Refactor CreateSchedule screen (split vc into separate views)
